+++ routes/api/authRouter.js
@@ -1,96 +1,9 @@
 const express = require("express");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-const Joi = require("joi");
-const User = require("../../models/user");
+const authController = require("../../controllers/authController");
 
 const router = express.Router();
 
-const registerSchema = Joi.object({
-  email: Joi.string().email().required(),
-  password: Joi.string().min(6).required(),
-});
-
-const loginSchema = Joi.object({
-  email: Joi.string().email().required(),
-  password: Joi.string().required(),
-});
-
-router.post("/users/register", async (req, res) => {
-  const { error } = registerSchema.validate(req.body);
-  if (error) {
-    return res.status(400).json({ message: error.details[0].message });
-  }
-
-  const { email, password } = req.body;
-
-  try {
-    const existingUser = await User.findOne({ email });
-
-    if (existingUser) {
-      return res.status(409).json({ message: "Email in use" });
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 10);
-
-    const newUser = new User({
-      email,
-      password: hashedPassword,
-      subscription: "starter",
-    });
-
-    await newUser.save();
-
-    const token = jwt.sign({ id: newUser._id }, process.env.USER_TOKEN, {
-      expiresIn: "1h",
-    });
-
-    res.status(201).json({
-      user: {
-        email: newUser.email,
-        subscription: newUser.subscription,
-      },
-      token,
-    });
-  } catch (error) {
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
-
-router.post("/users/login", async (req, res) => {
-  const { error } = loginSchema.validate(req.body);
-  if (error) {
-    return res.status(400).json({ message: error.details[0].message });
-  }
-
-  const { email, password } = req.body;
-
-  try {
-    const user = await User.findOne({ email });
-
-    if (!user) {
-      return res.status(401).json({ message: "Invalid credentials" });
-    }
-
-    const isPasswordValid = await bcrypt.compare(password, user.password);
-    if (!isPasswordValid) {
-      return res.status(401).json({ message: "Invalid credentials" });
-    }
-
-    const token = jwt.sign({ id: user._id }, process.env.USER_TOKEN, {
-      expiresIn: "1h",
-    });
-
-    res.json({
-      user: {
-        email: user.email,
-        subscription: user.subscription,
-      },
-      token,
-    });
-  } catch (error) {
-    res.status(500).json({ message: "Internal server error" });
-  }
-});
+router.post("/users/register", authController.registerUser);
+router.post("/users/login", authController.loginUser);
 
 module.exports = router;
